+++ src/react-components/src/__tests__/components/HelperErrorText-test.tsx
@@ -1,4 +1,3 @@
-
 import { expect } from "chai";
 import * as Enzyme from "enzyme";
 import * as React from "react";
@@ -10,6 +9,13 @@ describe("HelperErrorText Test", () => {
   it("Renders HelperErrorText", () => {
     const container = Enzyme.mount(<HelperErrorText id="helperTextWithLink"
     isError={false}>Text</HelperErrorText>);
+    expect(container.exists("#helperTextWithLink")).to.equal(true);
+  });
+
+  it("Has 'error' modifier if error is passed", () => {
+    const container = Enzyme.mount(<HelperErrorText id="helperTextWithLink"
+    isError={true}>Text</HelperErrorText>);
+    console.log("debug", container.debug());
     expect(container.exists("#helperTextWithLink")).to.equal(true);
   });
 });