+++ src/react-components/src/components/01-atoms/Button/Button.stories.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 
 import Button from "./Button";
 import { action } from "@storybook/addon-actions";
+import { ButtonTypes } from "./ButtonTypes";
 
 export default {
   title: "Button",
@@ -13,19 +14,21 @@ export const buttonFilledIcon = () =>
     callback={action("clicked")}
     id="button"
     content="search"
-    type="filled"
+    buttonType={ButtonTypes.Filled}
+    type="submit"
     iconPosition="left"
     iconName="search-small"
     iconDecorative={true}
-  />
-  ;
-export const buttonOutline = () => <Button callback={action("clicked")} id="button" content="Hello World" type="outline" />;
-export const buttonWithSpan = () => <Button callback={action("clicked")} id="button" type="outline" content={<span>Style my span!</span>} />;
+  />;
+
+export const buttonOutline = () => <Button callback={action("clicked")} id="button" content="Hello World" buttonType={ButtonTypes.Outline} />;
+export const buttonWithSpan = () => <Button callback={action("clicked")} id="button" buttonType={ButtonTypes.Outline} content={<span>Style my span!</span>} />;
 export const buttonIconChildrenString = () =>
   <Button
     callback={action("clicked")}
     id="button"
-    type="filled"
+    buttonType={ButtonTypes.Filled}
+    type="submit"
     iconPosition="right"
     iconName="search-small"
     iconDecorative={true}
@@ -35,7 +38,8 @@ export const buttonIconChildrenElement = () =>
   <Button
     callback={action("clicked")}
     id="button"
-    type="outline"
+    buttonType={ButtonTypes.Outline}
+    type="button"
     iconPosition="right"
     iconName="search-small"
     iconDecorative={true}
@@ -47,7 +51,8 @@ export const buttonIconRightLarge = () =>
   <Button
     callback={action("clicked")}
     id="button"
-    type="outline"
+    type="submit"
+    buttonType={ButtonTypes.Outline}
     iconPosition="right"
     modifiers={["large"]}
     iconName="plus"
@@ -60,7 +65,8 @@ export const linkLikeButton = () =>
   <Button
     callback={action["clicked"]}
     id="button"
-    type="link"
+    type="button"
+    buttonType={ButtonTypes.Link}
   >
     <span>button content</span>
   </Button >;
@@ -69,7 +75,8 @@ export const linkLikeButtonIconLeft = () =>
   <Button
     callback={action["clicked"]}
     id="button"
-    type="link"
+    type="button"
+    buttonType={ButtonTypes.Link}
     iconPosition="left"
     iconName="arrow-xsmall"
     iconModifiers={["left"]}