+++ src/react-components/src/components/01-atoms/Button/ButtonTypes.tsx
@@ -1,0 +1,5 @@
+export enum ButtonTypes {
+  Outline = "outline",
+  Filled = "filled",
+  Link = "link"
+}